certificate/requirements.txt
pandas==2.0.3
Pillow==9.4.0
Pillow==10.3.0
